fastapi_backend_version/requirements.txt
fastapi
boto3
uvicorn
python-dotenv
requests
celery
redis
sse-starlette
aioboto3
flpc